# Backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GitRank API",
    "version": "0.1.0",
    "description": "Tracks GitHub activity of members across cohorts, ranks them on leaderboards, and awards titles (records + badges). Every joiner is also auto-added to the always-on `global` cohort, which uses a rolling 365-day window. Reads are public; `/admin/*` routes require a static bearer token."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Static admin token from the `ADMIN_TOKEN` env var."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Cohort not found: devsoc-2025"
              },
              "details": {
                "type": "array",
                "description": "Present on validation errors.",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "Cohort": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "devsoc-2025"
          },
          "name": {
            "type": "string",
            "example": "DevSoc Training Program 2025"
          },
          "kind": {
            "type": "string",
            "enum": [
              "PROGRAM",
              "GLOBAL"
            ],
            "description": "PROGRAM cohorts are time-boxed training cohorts. The single GLOBAL cohort (slug `global`) covers every joiner across a rolling 365-day window."
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "memberCount": {
            "type": "integer",
            "description": "Present on list/detail responses."
          }
        }
      },
...
      "CohortDetail": {
        "type": "object",
        "properties": {
          "cohort": {
            "$ref": "#/components/schemas/Cohort"
          }
        }
      },
...
      "Snapshot": {
        "type": "object",
        "nullable": true,
        "description": "Latest stats for a member in a cohort; null if never synced.",
        "properties": {
          "capturedAt": {
            "type": "string",
            "format": "date-time"
          },
          "totalCommits": {
            "type": "integer"
          },
          "totalContributions": {
            "type": "integer"
          },
          "totalPRs": {
            "type": "integer"
          },
          "mergedPRs": {
            "type": "integer"
          },
          "reviewsGiven": {
            "type": "integer"
          },
          "issuesOpened": {
            "type": "integer"
          },
          "followers": {
            "type": "integer"
          },
          "totalStars": {
            "type": "integer"
          },
          "repoCount": {
            "type": "integer"
          },
          "contributedRepoCount": {
            "type": "integer"
          },
          "languageCount": {
            "type": "integer"
          },
          "topLanguages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "bytes": {
                  "type": "integer"
                }
              }
            }
          },
          "longestStreak": {
            "type": "integer"
          },
          "currentStreak": {
            "type": "integer"
          },
          "maxCommitsInOneDay": {
            "type": "integer"
          },
          "weekendCommitRatio": {
            "type": "number",
            "format": "float"
          },
          "nightCommitRatio": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        }
      },
...
      "AwardValue": {
        "type": "object",
        "additionalProperties": true,
        "description": "Title-specific value, e.g. `{ \"totalCommits\": 480 }`."
      },
...
      "MemberTitle": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "RECORD",
              "BADGE"
            ]
          },
          "flavor": {
            "type": "string",
            "nullable": true
          },
          "cohort": {
            "type": "object",
            "properties": {
              "slug": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "value": {
            "$ref": "#/components/schemas/AwardValue"
          },
          "awardedAt": {
            "type": "string",
            "format": "date-time"
          },
          "revokedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "MemberProfile": {
        "type": "object",
        "properties": {
          "member": {
            "type": "object",
            "properties": {
              "githubUsername": {
                "type": "string"
              },
              "zid": {
                "type": "string"
              },
              "displayName": {
                "type": "string",
                "nullable": true
              },
              "avatarUrl": {
                "type": "string",
                "nullable": true
              },
              "githubId": {
                "type": "integer",
                "nullable": true
              },
              "accountCreatedAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "cohorts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "cohort": {
                  "$ref": "#/components/schemas/Cohort"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "PARTICIPANT",
                    "ORGANISER"
                  ]
                },
                "joinedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "programRepos": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "owner": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                },
                "stats": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            }
          },
          "titles": {
            "type": "array",
            "description": "Record titles held (including past/revoked), with their cohort.",
            "items": {
              "$ref": "#/components/schemas/MemberTitle"
            }
          },
          "badges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberTitle"
            }
          }
        }
      },
...
      "CreateCohortRequest": {
        "type": "object",
        "required": [
          "name",
          "slug",
          "startDate"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "DevSoc Training Program 2025"
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "example": "devsoc-2025"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "SyncSummary": {
        "type": "object",
        "properties": {
          "sync": {
            "type": "object",
            "properties": {
              "cohortId": {
                "type": "string"
              },
              "cohortSlug": {
                "type": "string"
              },
              "membersSynced": {
                "type": "integer"
              },
              "snapshotsCreated": {
                "type": "integer"
              },
              "errors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "evaluation": {
            "type": "object",
            "properties": {
              "records": {
                "type": "integer"
              },
              "badges": {
                "type": "integer"
              },
              "members": {
                "type": "integer"
              }
            }
          }
        }
      },
      "DeleteMemberResult": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "string"
          },
          "reevaluatedCohorts": {
            "type": "integer"
          }
        }
      }
    }
  },
  "paths": {
...
    "/members/{username}": {
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "Member profile",
        "description": "Latest stats per cohort, record titles held (incl. past), and badges.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "octocat"
          }
        ],
        "responses": {
          "200": {
            "description": "Member profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberProfile"
                }
              }
            }
          },
          "404": {
            "description": "Unknown member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/cohorts": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create a cohort",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCohortRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created cohort",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CohortDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid admin token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate slug",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/members/{username}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Remove a member",
        "description": "Cascade-deletes memberships, snapshots and awards, then re-evaluates affected cohorts so vacated records transfer to the runner-up.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteMemberResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid admin token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/sync/{slug}": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Trigger a manual sync + title evaluation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cohort slug (kebab-case).",
            "example": "devsoc-2025"
          }
        ],
        "responses": {
          "200": {
            "description": "Sync + evaluation summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncSummary"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid admin token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown slug",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/sync-all": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Run the sync + eval runner across every active cohort",
        "description": "External-cron trigger for free-tier hosts. Runs the same in-process runner as `node-cron` and shares its lock: if a sync is already running, returns `{ \"skipped\": true }` with 200 instead of double-syncing.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Runner summary (or `{ skipped: true }` if already running).",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "skipped": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        },
                        "startedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "finishedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "summaries": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "cohortId": {
                                "type": "string"
                              },
                              "cohortSlug": {
                                "type": "string"
                              },
                              "membersSynced": {
                                "type": "integer"
                              },
                              "snapshotsCreated": {
                                "type": "integer"
                              },
                              "errors": {
                                "type": "array",
                                "items": {
                                  "type": "object"
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "skipped": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid admin token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}